Multi-Crate and Structured CLI
Very modular setup with multiple crates in the Cargo workspace, for now
just the CLI and web crates, but more to come. Structured subcommands
exist now for generating JWK keys `generate jwk` and for launching
rocket `serve`. With this setup, we could also add a `rswa-frontend` in
Yew or another Rust front-end framework.